refactor: split lib/provider.ts into message-formatting and response-parsing
